// src/shared/domain/enums/tms-module.enum.ts
export enum TmsModule {
  TRIPS = 'TRIPS',
  VEHICLES = 'VEHICLES',
  DRIVERS = 'DRIVERS',
  APPOINTMENTS = 'APPOINTMENTS',
  CHECKLIST = 'CHECKLIST',
  CARRIERS = 'CARRIERS',
  DOMICILIARY = 'DOMICILIARY',
  RATES = 'RATES',
  CREDIT_NOTES = 'CREDIT_NOTES',
  USERS = 'USERS',
  AUDIT = 'AUDIT',
  ROUTE_POINTS = 'ROUTE_POINTS',
  WAREHOUSES = 'WAREHOUSES',
}